source/requirements/base.txt
bumpversion==0.5.3
coreapi==2.3.3
Django==2.2.3
djangorestframework==3.10.1
djangorestframework-gis==0.14
django-filter==2.2.0
django-oauth-toolkit==1.2.0
luizalabs-django-toolkit==2.1.0
Markdown==3.1.1
whitenoise==4.1.2